updated Kernels
Added a bare metal (Core) Kernel for use with custom Kernels
moved testing registration into the testing Kernel
lazy load Kernel imports as they  may not be required if custom Kernel is used
